requirements.txt
requests
beautifulsoup4
Fabric3
pyTelegramBotAPI
sqlalchemy
